% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear_selectivity.R
\name{AdjustedFreq}
\alias{AdjustedFreq}
\alias{AdjustedFreq.default}
\alias{AdjustedFreq.RelativeRetention}
\title{Get adjusted frequency based on relative retention results.}
\usage{
AdjustedFreq(x, ...)

\method{AdjustedFreq}{default}(x, ...)

\method{AdjustedFreq}{RelativeRetention}(x, ...)
}
\arguments{
\item{x}{Generic for S3 methods.}

\item{...}{Passed to other methods.}
}
\value{
A dataframe of catch and adjusted catch per length (row).
}
\description{
Provides a convenient way to adjust catch overall given
  results of gear selectivity (i.e., \code{NetFet} model results).

S3 method for class \code{RelativeRetention}.
}
\examples{
# code here
}
