% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear_selectivity.R
\name{ApplyNetFit}
\alias{ApplyNetFit}
\title{Apply Millar's \code{NetFit} to catch data using gill net or trammel net.}
\usage{
ApplyNetFit(data, len, mesh, meshUnit = "in", relPower = NULL)
}
\arguments{
\item{data}{A dataframe with at least length and net mesh (size) fields.}

\item{len}{Field name within data containing (numeric) length measurement.}

\item{mesh}{Field name within data containing size of net mesh. Field data
can be character but must be "coerceable" to numeric.}

\item{meshUnit}{Scalar character containing unit of measure of net. Choices
are "in" (inches), "ft" (feet), "mm" (millimeters), "m" (meter), and
"cm" (centimeters). \code{ApplyNetFit} will convert to cm.}

\item{relPower}{The fishing power of each net size. Leave NULL (default)
if only single mesh or all meshes are fished equally. Otherwise, supply
numeric vector (length = number of mesh sizes) representing net
configuration effort. For example, if net mesh sizes are 6", 7", & 8" and
the 8" mesh is fished at twice the effort, relPower = c(1, 1, 2) [for
6", 7", & 8"].}
}
\value{
A list of model fits (see details).
}
\description{
Apply Millar's \code{NetFit} to catch data using gill net or trammel net.
}
\details{
\code{ApplyNetFit} fits 5 Millar models: \code{norm.loc};
   \code{norm.sca}; \code{lognorm}; \code{binorm.sca}; and \code{bilognorm}.
   Each one of these models is a list with 12 items (explained below).
}
\section{Value}{


\code{ApplyNetFit} returns a list of class "NetFit" each list item
   containing the components below.

\code{par}         see **Value** in \code{\link{optim}}

\code{value}       see **Value** in \code{\link{optim}}

\code{counts}      see **Value** in \code{\link{optim}}

\code{convergence} see **Value** in \code{\link{optim}}

\code{message}     see **Value** in \code{\link{optim}}

\code{hessian}     see **Value** in \code{\link{optim}}

\code{Deviance}    model deviance (value)

\code{deviance}    model deviance (function); not sure of purpose

\code{rtype}       the name (as character) of the model

\code{rel.power}   values supplied to relPower, relative fishing power
                   of the gear

\code{Meshsize}    mesh size (in centimeters)

\code{Data}        length frequency by mesh size (raw count data)
}

\examples{
ApplyNetFit(
  data = trammel_catch,
  len = FL,
  mesh = MeshSize,
  meshUnit = "in",
  relPower = c(1, 1, 2)
)
}
\references{
https://www.stat.auckland.ac.nz/~millar/selectware/RNext/

            https://www.stat.auckland.ac.nz/people/rmil013

            Millar and Holst (1997). Estimation of gillnet and hook
            selectivity using log-linear models
            ICES J. Mar. Sci. 54: 471-477.
}
