% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear_selectivity.R
\name{DeviancePlots}
\alias{DeviancePlots}
\alias{DeviancePlots.default}
\alias{DeviancePlots.NetFit}
\title{Generate Millar's deviance plots for gear selectivity.}
\usage{
DeviancePlots(x, ...)

\method{DeviancePlots}{default}(x, ...)

\method{DeviancePlots}{NetFit}(x, ...)
}
\arguments{
\item{x}{Generic parameter for S3 method. Likely a model fit
of class "NetFit".}

\item{...}{Passed to other methods.}
}
\value{
If \code{x} is a list, output is a list with same number of
  elements, each element is a matrix. Otherwise output is a matrix with one
  column. See Values.
}
\description{
We choose a selectivity model based on deviance. The lower the
  deviance, the better the fit. Deviance plots display residuals (by mesh
  size) in a "bubble-plot" format. (describe 'bubbles' and size here) Plot is
  a side effect of Millar's \code{Summary} function. Function returns a
  list, with number of items equal to the number of models (i.e., 5). See
  Values for output.

Default S3 method.

S3 method for class NetFit.
}
\note{
\code{DeviancePlots} is essentially a convenient wrapper for Millar's
   \code{Summary} function.
}
\section{Values}{


\describe{
  \item{null.l}{doc needed here}
  \item{model.l}{doc needed here}
  \item{full.l}{doc needed here}
  \item{Deviance}{model deviance}
  \item{Pearson.chisq}{doc needed here}
  \item{d.o.f}{degrees of freedom}
}
}

\examples{
# for method .NetFit
net_fit <- ApplyNetFit(
  data = trammel_catch,
  len = FL,
  mesh = MeshSize,
  meshUnit = "in",
  relPower = c(1,1,2)
)
DeviancePlots(net_fit)
}
