% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_functions.R
\name{input}
\alias{input}
\alias{SpawningProb}
\alias{EggCount}
\alias{AgeDist}
\alias{SurvivalProb}
\alias{CreateSPDataframe}
\title{Suite of functions to generate model data inputs.}
\usage{
SpawningProb(pMat = NULL, len, age, mature = 0.15)

EggCount(numEggs = NULL, len, age)

AgeDist(ageFreq, abund = 48000, fracFemale = 0.5)

SurvivalProb(
  ...,
  mu = 0.01,
  agesMu = 10:15,
  estS = 0.8132,
  estMu = 0.1364,
  methodSB = TRUE
)

CreateSPDataframe(ages, sRate, sRateErr)
}
\arguments{
\item{pMat}{A dataframe with two fields: length & probability of maturity
for each length. Default uses internal \code{SpawningProbWST}
(for White Sturgeon) & from Chapman 1989 (?).}

\item{len}{Numeric vector of lengths supplied as new data to \code{predict}.}

\item{age}{Numeric vector of ages, where each age is appropriate for given
length supplied in \code{len}. Vector must be same length as \code{len}.}

\item{mature}{Fraction of females spawning in a given year. Default 0.15.}

\item{numEggs}{A dataframe with two fields: length & number of eggs.
for each length. Default uses internal \code{FecundityWST}
(for White Sturgeon) & from Devore 19??, using \eqn{0.072 * len^2.94}.}

\item{ageFreq}{A numeric vector of age frequency (like that derived from
age-length key or direct ageing).}

\item{abund}{A numeric scalar given estimated overall abundance for species
being modeled. Default 48,000 (as estimated by CDFW).}

\item{fracFemale}{A numerice scalar given fraction of females in population.
Default 0.5 from Chapman et al. 1996.}

\item{...}{Passed to internal function \code{CreateSPDataframe}.}

\item{mu}{A numeric vector giving exploitation (harvest rate;
between 0 and 1). Default 0.01.}

\item{agesMu}{A numeric vector of ages on which exploitation applies
(i.e., age group subject to harvest). Default 10:15.}

\item{estS}{A numeric scalar given estimated survival rate of (typically)
adult fish. Derived using \code{ChapmanRobson}. Default 0.8132 from
S. Blackburn thesis work.}

\item{estMu}{A numeric scalar given estimated exploitation rate of
harvestable fish. Derived using \code{Exploitation}. Default 0.1364 from
S. Blackburn thesis work using 2014-2016 CDFW mark-recapture data.}

\item{methodSB}{Logical. Default (TRUE) denotes using S. Blackburn method to
calculate \code{F} using \code{Z} & \code{A} calculated from \code{estS}
and not from \code{S0}. Change to FALSE to use \code{S0}.}

\item{ages}{A numeric vector of ages (e.g., 0:19).}

\item{sRate}{A numeric vector giving survival rate (between 0 and 1).}

\item{sRateErr}{A numeric vector given standard error of \code{sRate}.}
}
\description{
\code{spopmodel} requires specific inputs to operate. In the
   absence of such inputs, functions herein can generate dataframes to
   satisfy model requirements. If such data already exists, however, then
   it's best to use those data. NOTE: Functions herein default for use with
   San Francisco Estuary-based White Sturgeon. So do keep that in mind when
   employing these functions.

Uses \code{glm} to predict spawning probability
   based on probability of maturity at length \emph{i}. Spawning probability
   is then adjusted for fraction of females spawning annually.

Uses \code{lm} to predict number of eggs given
   length \emph{i}.

Uses age frequency to provide by age (1) estimated
   abundance & (2) estimated abundance of females. Uses arguments supplied
   to \code{abund} & \code{fracFemale} to achieve this.
}
\note{
Ideally \code{ages} vector supplied from age distribution data (e.g.,
   \code{age_dist}). Vector should contain age-0 sequentially through oldest
   age. \code{sRate} and \code{sRateErr} need not match \code{ages} in
   length, just not longer. If less than, the last value (e.g.,
   \code{sRate[length(sRate)]} is repeated to complete the dataframe.
   \code{sRate[length(sRate)]} must be S0, that is survival rate considering
   only natural mortality.
}
\section{\code{SpawningProb}}{
NA
}

\section{\code{EggCount}}{
NA
}

\section{\code{AgeDist}}{
NA
}

\section{\code{SurvivalProb}}{
NA
}

\section{\code{CreateSPDataframe}}{
NA
}

\keyword{internal}
