% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{Lambda}
\alias{Lambda}
\title{Calculate Mean Population Growth Rate.}
\usage{
Lambda(popChanges, selectCI = NULL)
}
\arguments{
\item{popChanges}{A numeric vector providing population changes. Likely the
output from \code{PopProjections}.}

\item{selectCI}{Numeric value of either 0.90, 0.95, or 0.99. Denotes
confidence interval (to \code{probs} argument) using sample quantiles
\code{\link[stats]{quantile}}. Default is NULL, which is the same as 0.95.}
}
\value{
A dataframe with seven fields, including mean lambda.
}
\description{
Given population changes over specified period, \code{Lambda}
  calculates mean lambda (population growth rate) with lower & upper bounds.
  Lambda = 1 indicates a stable population.
}
\examples{
# coming soon
}
