% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear_selectivity.R
\name{LoadMillarCode}
\alias{LoadMillarCode}
\title{Download R. Millar gear selectivity code.}
\usage{
LoadMillarCode()
}
\value{
Message giving file path to where files were downloaded.
}
\description{
\code{LoadMillarCode} downloads to user's computer
   gear selectivity source code written by Russel Millar. Files
   downloaded currently are NextGeneration.R & SelnCurveDefinitions.R,
   the main two used in this package. Files download to root directory.
}
\note{
Only run this code if interested in seeing source files. Other-
   wise, gear selectvity functions within this package will suffice.
}
\examples{
\dontrun{
LoadMillarCode()
}
}
\references{
https://www.stat.auckland.ac.nz/~millar/selectware/RNext/

            https://www.stat.auckland.ac.nz/people/rmil013

            Millar and Holst (1997). Estimation of gillnet and hook
            selectivity using log-linear models
            ICES J. Mar. Sci. 54: 471-477.
}
