% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{plot.Frequency}
\alias{plot.Frequency}
\title{Generate base R histogram from object of class \code{\link{Frequency}}.}
\usage{
\method{plot}{Frequency}(x, ...)
}
\arguments{
\item{x}{An object of class Frequency (& histogram).}

\item{...}{Passed on to other methods.}
}
\value{
Displays a histogram using frequency item in \code{x}.
}
\description{
Plot a histogram using \code{hist} (esentially
   calling next method on \code{plot}, which calls
   \code{plot.histogram}). Display contains sample size along
   with some descriptive statistics.
}
\examples{
freq <- Frequency(trammel_catch[["FL"]], binWidth = 5)
plot(freq)
}
