% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear_selectivity.R
\name{plot.RelativeRetention}
\alias{plot.RelativeRetention}
\title{Plot relative retention values for each gear size.}
\usage{
\method{plot}{RelativeRetention}(x, ...)
}
\arguments{
\item{x}{Object of class \code{RelativeRetention}.}

\item{...}{Passed on to other methods.}
}
\value{
Nothing. But does display (base R) line plot of relative renetions
   as a function of measured length (e.g., fork length). Plot title
   contains model name, model deviance, and true or false for
   "is standardized?". See default method
   \code{\link{RelativeRetention}}.
}
\description{
This is an S3 method for \code{plot} of class
   \code{RelativeRetention}. It allows the user to visually
   display relative retentions of each gear size and to
   view on each plot model deviance (for ease of deciding best fit).
}
\examples{
net_fit <- ApplyNetFit(
  data = trammel_catch,
  len = FL,
  mesh = MeshSize,
  meshUnit = "in",
  relPower = c(1,1,2)
)
rr <- RelativeRetention(net_fit)
plot(rr$norm.loc)
}
