% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{PopProjections}
\alias{PopProjections}
\title{Get population projection over defined period (years).}
\usage{
PopProjections(fSims, sSims, mn, sdev, ageFreq, period)
}
\arguments{
\item{fSims}{A matrix of fecundity simulations, likely the product of
simulated number of eggs x simulated spawning probabilities x sex
ratio (expressed as a proportion, e.g., 0.5).}

\item{sSims}{A matrix of survival simulations, likely one matrix for each
level of exploitation (mu) being tested.}

\item{mn}{Scalar numeric value as survival probability of oldest fish (i.e.,
max age in \code{ageFreq})}

\item{sdev}{Scalar numeric value as corresponding error for \code{mn}.}

\item{ageFreq}{Numeric vector of age frequency passed on to \code{n} in
\code{\link{pop.projection}}.}

\item{period}{Scalar integer given number of years to project. Passed on to
\code{iterations} in \code{\link{pop.projection}}.}
}
\value{
A large matrix with rows as 'Value' output in
  \code{link{pop.projection}} and number of columns = number of simulations.
}
\description{
\code{PopProjections} is essentially a wrapper for
  \code{\link{pop.projection}}. It will run \code{pop.projection} for every
  simulation (i.e., column) in the simulation matrices (passed to
  \code{fSims} & \code{sSims}). Function creates the Leslie Matrix (*A*)
  required by \code{pop.projection}.
}
\examples{
# code here
}
