% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gear_selectivity.R
\name{RelativeRetention}
\alias{RelativeRetention}
\alias{RelativeRetention.default}
\alias{RelativeRetention.NetFit}
\title{Get gear relative retention based on net fit model output.}
\usage{
RelativeRetention(x, ...)

\method{RelativeRetention}{default}(x, ..., standardize = TRUE)

\method{RelativeRetention}{NetFit}(x, ...)
}
\arguments{
\item{x}{Generic for S3 methods. Model output from \code{NetFit}.}

\item{...}{Passed to other methods.}

\item{standardize}{Boolean. If \code{TRUE} (default), relative
retention values are divided by \code{max(relative retention)}.
If \code{FALSE}, not such division occurs.}
}
\value{
List of objects. See Values.
}
\description{
\code{RelativeRetention} is a re-write of Millar's
   \code{PlotCurves} for convenient use within this package. It
   generates (among other items) a dataframe relative retention values
   for each gear type (columns) by each length (or length bin; rows).
   Out put can then be used to 'adjust' catch accordingly.

Default S3 method.

S3 method for class \code{NetFit}.
}
\section{Values}{


\describe{
  \item{Data}{dataframe of relative retention values}
  \item{Freq}{raw frequency data (i.e., recorded catch)}
  \item{RelPower}{relative power of gear type}
  \item{MeshSize}{mesh size of each gear}
  \item{Deviance}{measure of 'fit' for each model}
  \item{ColNames}{column names of \code{Data}}
  \item{Model}{name of net fit model}
  \item{Standardized}{true or false, is relative retention standardized}
}
}

\examples{
# for .NetFit method
net_fit <- ApplyNetFit(
  data = trammel_catch,
  len = FL,
  mesh = MeshSize,
  meshUnit = "in",
  relPower = c(1,1,2)
)
RelativeRetention(net_fit)
}
