% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_functions.R
\name{Simulations}
\alias{Simulations}
\title{Generate simulations for stochastic modeling of population trajectory.}
\usage{
Simulations(
  data,
  prob,
  std,
  recruitment = 5,
  minb = 0,
  maxb = NULL,
  iters = 10,
  seed = NULL,
  type = c("survival", "spawning", "numeggs")
)
}
\arguments{
\item{data}{Dataframe containing age-specific probabilities for survival,
spawning, or number of eggs along with standard deviations (errors)}

\item{prob}{Column name in data containing age-specific probabilities or
count.}

\item{std}{Column name in data containing error of probabilities.}

\item{recruitment}{Used only for "survival" option. Value (default = 5)
denotes period of successful recruitment.}

\item{minb}{Used only for "numeggs" option. Value passed on to
\code{\link{stretchbetaval}}; default = 0.}

\item{maxb}{Used only for "numeggs" option. Value passed on to
\code{\link{stretchbetaval}}; default = NULL; if NULL value will be 3 times
data[[prob]]}

\item{iters}{Scalar integer. Number of iterations to perform. Default 10;
function will slow with higher (i.e., > 5000) value.}

\item{seed}{Numeric value passed to \code{set.seed} to ensure repeatability.}

\item{type}{Scalar character as one of three options: "survival"; "spawning";
or "numeggs". Choose one based on desired simulation. Default = "survival".}
}
\value{
A matrix of model simulations (varies depending on argument
   supplied to \code{type}). Matrix is of class \code{simulations}.
}
\description{
Simulate environmental stochasticity for surivial, spawning, and
  number of eggs using age-specific empirical data. Simulated data are used
  by \code{PopProjections} to calculate population trajections (lambda).
}
\examples{
# code here
}
