% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/age_length.R, R/frequency.R,
%   R/gear_selectivity.R, R/growth.R, R/millar_curvedefs.R,
%   R/millar_selectivity.R, R/model_functions.R, R/model_sims.R,
%   R/spopmodel-internals.R, R/survival.R, R/utility_functions.R
\name{CheckAge}
\alias{CheckAge}
\alias{BinWidth}
\alias{StartVals}
\alias{GetX0}
\alias{print.RelativeRetention}
\alias{GetVBGM}
\alias{selncurves}
\alias{NetFit}
\alias{nllhood}
\alias{Estimates}
\alias{PlotCurves}
\alias{Summary}
\alias{LeslieMatrix}
\alias{MeanLogLambda}
\alias{bval}
\alias{sbval}
\alias{pp}
\alias{BetavalSims}
\alias{SBetavalSims}
\alias{SampleSims}
\alias{spopmodel-internals}
\alias{FishingParams}
\alias{ToCM}
\alias{CreateLenBins}
\alias{CheckBin}
\title{Internal functions for \code{spopmodel}.}
\usage{
CheckAge(age)

BinWidth(w)

StartVals(len, mesh, FUN = NULL)

GetX0(svals)

\method{print}{RelativeRetention}(x, ...)

GetVBGM(Linf, K, t0)

selncurves(rtype)

NetFit(Data, Meshsize, x0, rtype = "norm.loc", rel.power = NULL)

nllhood(theta, Data, Meshsize, r, rel.power)

Estimates(fit)

PlotCurves(fit, Meshsize = NULL, plotlens = NULL, standardize = TRUE, ...)

Summary(
  fit,
  label = "Deviance residuals",
  xlabel = "Length (cm)",
  ylabel = "Mesh size (cm)",
  cex = 1
)

LeslieMatrix(sims, survSims, mn, sdev)

MeanLogLambda(popChanges)

bval(mn, sdev, fx = runif(1))

sbval(mn, std, minb, maxb, fx)

pp(A, n, iterations = 20)

BetavalSims(mn, sdev, iterations = 10)

SBetavalSims(mn, sdev, minb = 0, maxb = NULL, iterations = 10)

SampleSims(sims, n, span = 5, seed = NULL)

FishingParams(S, mu)

ToCM(x, unit = c("in", "ft", "mm", "m", "cm"))

CreateLenBins(len, lenBreaks, breakLabs = NULL, numericBin = FALSE, ...)

CheckBin(x)
}
\description{
Functions herein are used internally by \code{spopmodel}.
   Direct function calls are not required by general user.
}
\keyword{internal}
